backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6

# PowerPoint generation
python-pptx>=0.6.21

# Azure SDK
azure-cosmos>=4.5.0
azure-storage-blob>=12.18.0
azure-identity>=1.15.0
azure-ai-projects>=1.0.0b3

# A2A SDK
a2a-python-sdk>=0.1.0

# Semantic Kernel
semantic-kernel>=0.9.0

# HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# OpenTelemetry
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-auto-instrumentation>=0.42b0
opentelemetry-exporter-otlp>=1.21.0

# Image processing
Pillow>=10.0.0
requests>=2.31.0

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.6.0